rename quantum_test_base to common
