+++ azure-quantum/tests/unit/test_job.py
@@ -19,7 +19,7 @@ from azure.quantum import Workspace
 from azure.quantum.optimization import Problem
 from azure.quantum.optimization.solvers import SimulatedAnnealing
 from azure.quantum import Job
-from ..quantum_test_base import QuantumTestBase
+from ..common import QuantumTestBase
 
 
 class TestJob(QuantumTestBase):
