+++ azure-quantum/tests/unit/test_optimization.py
@@ -14,7 +14,7 @@ import unittest
 from azure.quantum import Workspace
 from azure.quantum.optimization import Problem, ProblemType, Term
 from azure.quantum.optimization.solvers import ParallelTempering, SimulatedAnnealing, HardwarePlatform, QuantumMonteCarlo
-from ..quantum_test_base import QuantumTestBase 
+from ..common import QuantumTestBase 
 
 class TestProblem(QuantumTestBase):
     def test_add_terms(self):
